Complete day 11

+++ src/de/ungefroren/adventofcode/y2019/intcodeComputer/IntcodeComputer.scala
@@ -171,5 +171,5 @@ object IntcodeComputer {
     new IntcodeComputer(memory)(instructions)
 
   def apply(source: Source)(implicit instructions: Set[Instruction]): IntcodeComputer =
-    new IntcodeComputer(source.mkString.replaceAll("\\s", "").split(",").map(_.toInt))(instructions)
+    new IntcodeComputer(source.mkString.replaceAll("\\s", "").split(",").map(_.toLong))(instructions)
 }
